+++ src/main/java/com/github/space125/jrtb/jrclient/JavaRushGroupClientImpl.java
@@ -3,12 +3,16 @@ package com.github.space125.jrtb.jrclient;
 import com.github.space125.jrtb.jrclient.dto.GroupDiscussionInfo;
 import com.github.space125.jrtb.jrclient.dto.GroupInfo;
 import com.github.space125.jrtb.jrclient.dto.GroupRequestArgs;
+import com.github.space125.jrtb.jrclient.dto.PostInfo;
 import kong.unirest.GenericType;
 import kong.unirest.Unirest;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Optional;
+
+import static org.springframework.util.CollectionUtils.isEmpty;
 
 /**
  * Implementation of the {@link JavaRushGroupClient} interface.
@@ -19,9 +23,11 @@ import java.util.List;
 public class JavaRushGroupClientImpl implements JavaRushGroupClient {
 
     private final String javarushApiGroupPath;
+    private final String javarushApiPostPath;
 
-    public JavaRushGroupClientImpl(@Value("${javarush.api.path}") String javarushApiGroupPath) {
-        this.javarushApiGroupPath = javarushApiGroupPath + "/groups";
+    public JavaRushGroupClientImpl(@Value("${javarush.api.path}") String javarushApiPath) {
+        this.javarushApiGroupPath = javarushApiPath + "/groups";
+        this.javarushApiPostPath = javarushApiPath + "/posts";
     }
 
     @Override
@@ -37,14 +43,15 @@ public class JavaRushGroupClientImpl implements JavaRushGroupClient {
     public List<GroupDiscussionInfo> getGroupDiscussionList(GroupRequestArgs args) {
         return Unirest.get(javarushApiGroupPath)
                 .queryString(args.populateQueries())
-                .asObject(new GenericType<List<GroupDiscussionInfo>>() {})
+                .asObject(new GenericType<List<GroupDiscussionInfo>>() {
+                })
                 .getBody();
     }
 
     @Override
     public Integer getGroupCount(GroupRequestArgs args) {
         return Integer.valueOf(
-                Unirest.get(String.format("%s/count",javarushApiGroupPath))
+                Unirest.get(String.format("%s/count", javarushApiGroupPath))
                         .queryString(args.populateQueries())
                         .asString()
                         .getBody()
@@ -56,5 +63,16 @@ public class JavaRushGroupClientImpl implements JavaRushGroupClient {
         return Unirest.get(String.format("%s/group%s", javarushApiGroupPath, id.toString()))
                 .asObject(GroupDiscussionInfo.class)
                 .getBody();
+    }
+
+    @Override
+    public Integer findLastArticleId(Integer groupSubId) {
+        List<PostInfo> posts = Unirest.get(javarushApiPostPath)
+                .queryString("order", "NEW")
+                .queryString("groupKid", groupSubId.toString())
+                .queryString("limit", 1)
+                .asObject(new GenericType<List<PostInfo>>() {
+                }).getBody();
+        return isEmpty(posts) ? 0 : Optional.ofNullable(posts.get(0)).map(PostInfo::getId).orElse(0);
     }
 }