+++ src/main/java/com/github/space125/jrtb/jrclient/dto/MeGroupInfo.java
@@ -9,6 +9,7 @@ import lombok.Data;
  */
 @Data
 public class MeGroupInfo {
+
     private MeGroupInfoStatus status;
     private Integer userGroupId;
 }