+++ src/main/java/com/github/space125/jrtb/jrclient/dto/PostInfoType.java
@@ -1,0 +1,10 @@
+package com.github.space125.jrtb.jrclient.dto;
+
+/**
+ * Enum which represents Post info type.
+ *
+ * @author Ivan Kurilov on 03.08.2021
+ */
+public enum PostInfoType {
+    UNKNOWN, USUAL, INNER_LINK, OUTER_LINK
+}