+++ src/main/java/com/github/space125/jrtb/jrclient/dto/UserDiscussionInfo.java
@@ -9,6 +9,7 @@ import lombok.Data;
  */
 @Data
 public class UserDiscussionInfo {
+
     private Boolean isBookmarked;
     private Integer lastTime;
     private Integer newCommentsCount;