+++ src/main/java/com/github/space125/jrtb/service/SendBotMessageServiceImpl.java
@@ -6,6 +6,11 @@ import org.springframework.stereotype.Service;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
+import java.util.List;
+
+import static org.springframework.util.CollectionUtils.isEmpty;
+
+
 /**
  * Implementation of {@link SendBotMessageService} interface.
  *
@@ -29,5 +34,12 @@ public class SendBotMessageServiceImpl implements SendBotMessageService {
         } catch (TelegramApiException e) {
             e.printStackTrace();
         }
+    }
+
+    @Override
+    public void sendMessage(String chatId, List<String> messages) {
+        if (isEmpty(messages)) return;
+
+        messages.forEach(message -> sendMessage(chatId, message));
     }
 }