+++ src/test/java/com/github/space125/jrtb/command/StartCommandTest.java
@@ -6,7 +6,8 @@ import org.junit.jupiter.api.DisplayName;
  * @author Ivan Kurilov on 10.07.2021
  */
 @DisplayName("Unit-level testing for StartCommand")
-public class StartCommandTest extends AbstractCommandTest{
+class StartCommandTest extends AbstractCommandTest{
+
     @Override
     String getCommandName() {
         return CommandName.START.getCommandName();