+++ src/test/resources/sql/clearDbs.sql
@@ -1,1 +1,3 @@
+delete from group_x_user;
+delete from group_sub;
 delete from tg_user;